+++ infra/bootstrapper/prod/providers.tf
@@ -1,0 +1,34 @@
+terraform {
+  required_providers {
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = "~> 4.0"
+    }
+
+    azuread = {
+      source  = "hashicorp/azuread"
+      version = "~> 2.0"
+    }
+
+    github = {
+      source  = "integrations/github"
+      version = "~> 6.0"
+    }
+  }
+
+  backend "azurerm" {
+    resource_group_name  = "dx-p-itn-tfstate-rg-01"
+    storage_account_name = "dxpitntfstatest01"
+    container_name       = "terraform-state"
+    key                  = "dx.bootstrapper.prod.tfstate"
+  }
+}
+
+# Default provider configuration
+provider "azurerm" {
+  features {}
+}
+
+provider "github" {
+  owner = "pagopa"
+}