+++ infra/modules/azure_service_bus_namespace/main.tf
@@ -6,7 +6,7 @@ terraform {
     }
     dx = {
       source  = "pagopa-dx/azure"
-      version = ">= 0.2.0"
+      version = "~> 0.3"
     }
   }
 }