+++ infra/modules/azure_core_values_exporter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # azure_core_values_exporter
 
+## 0.1.0
+
+### Minor Changes
+
+- c93e73f: Add application insights resource
+
+## 0.0.2
+
+### Patch Changes
+
+- a08a2c9: Update some configuration with new optional variables
+
 ## 0.0.1
 
 ### Patch Changes